Moved Navbar and Footer jsx files to layout folder and refactored
