+++ client/src/layout/Layout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Navbar from "../components/Navbar.jsx";
-import Footer from "../components/Footer.jsx";
+import Navbar from "./Navbar.jsx";
+import Footer from "./Footer.jsx";
 import { Outlet } from "react-router-dom";
 
 const Layout = () => {
